tp2/pyproject.toml
[project]
name = "compresor-imagenes-evolutivo"
version = "0.1.0"
description = "Aproximacion visual de imagenes mediante algoritmos geneticos y triangulos traslucidos."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "scikit-image>=0.21.0",
]

[project.optional-dependencies]
gpu = [
    "moderngl>=5.10.0",
]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.uv]
package = false
